Fix file selector bug
If the file selector is called with a non-existent path, an alert is
issued (as expected) but dismissing the alert leaves the path
unchanged, and an empty list is displayed.  This update changes the
path to the root directory of the current drive, which should always
be valid.  Thanks to Vincent Rivière for reporting this.