Fix display bug in EmuDesk initialisation
The menu bar was enabled too early, before the desktop had been
built.  If the user moved the mouse cursor into the menu bar very
quickly, a menu could be dropped down while the desktop display was
being built.  This caused incorrect background data to be saved, so
when it was restored, the desktop display was messed up.
This bug was present in the original imported source; thanks to
Vincent Rivière for reporting it.