Make EmuDesk icon/text selection like Atari TOS
Previously, 'display as text'/'display as icons' was a single menu
item with text that changed.  This was the only such menu item,
and only displayed the current selection indirectly.  There is now
one menu item for each choice, and a checkmark indicates the
currently-selected item, just like the sort sequence selection.