Reset slider when closing subfolder
This makes the display of the parent folder start with the first
item in the folder, just like TOS (and desktops on other OSs).
Previously the slider position was unchanged on a close, and the
display started at the same line as the subfolder.
Thanks to Thorsten Otto for pointing this out.