Improve TT RAM size detection for Storm cards
The Storm TT RAM card can have SIMMs of different sizes in its two
slots.  In this situation, Atari TOS sizes TT RAM correctly, while
EmuTOS doesn't, leading to crashes.  This occurs because EmuTOS
expects that accessing non-existent memory will always cause a bus
error which is not true for this configuration.  Now, if a bus error
does not occur, we write/read the memory to verify its presence.