Fix media change bug with write-protected floppies
For write-protected diskettes, flop_mediach() returned MEDIAMAYCHANGE
in many cases.  This caused Rwabs() to return E_CHNG, which caused
the diskette to be logged in each time it was accessed.  As a result,
EmuCON was effectively unable to change the default directory on a
write-protected diskette.  We now check the contents of the boot
sector & FAT1 for changes and return a definite MEDIACHANGE or
MEDIANOCHANGE instead.  Thanks to Vincent Riviere & Christian Zietz
for their help with this.