Fix bug in AES screen initialisation
The AES uses global variables to track the current values of some VDI
attributes, in order to avoid unnecessary VDI calls.  Some of these
were not reinitialised on resolution change, which could result in
some VDI attributes not being set, because the global indicated that
they were already set.
Specifically, when the resolution was changed from TT Low (or ST Low)
to TT Medium, the desktop icon text was drawn in the wrong font size.
This bug was in the original imported code, but was hidden until
commit 3795833, which changed the desktop initialisation sequence.