Fix bug in rsconf handling for SCC
The serial port code uses a global pointer to point to the IOREC
associated with the currently-mapped serial port.  This pointer is
updated by a Bconmap() call.  The old code erroneously used this
pointer in rsconf_scc() to set/retrieve the serial port parameters.
This does not work when rsconf_scc() is called via the pointer in
the MAPTAB array, since the global pointer will point to the last
Bconmap'd device.  We now pass rsconf_scc() a pointer to the
correct IOREC.